+++ tests/book.rs
@@ -1,5 +1,11 @@
 use file_format::FileFormat;
 
+#[test]
+fn test_broad_band_ebook() {
+    let fmt = FileFormat::from_file("fixtures/book/sample.lrf").unwrap();
+    assert_eq!(fmt, FileFormat::BroadBandEbook);
+}
+
 #[cfg(feature = "reader-zip")]
 #[test]
 fn test_electronic_publication() {